requirements.txt
# Dependências do projeto
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0

# Supabase
supabase==2.0.0
psycopg2-binary==2.9.7
sqlalchemy==2.0.21
flask-sqlalchemy==3.0.5

# Para desenvolvimento
pytest==7.4.2
black==23.7.0
flake8==6.0.0